outdooer-backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1